Added wishlist and closing operations
the closing operations can now be added to through the WindowListener added to the frame.
The wishlist can now be displayed for any specified user by parameter. The wishlist button has been added for the items being displayed
The images are for the wishlist button, for on and off the wishlist